Load the WebIDL files specified in the binding
This loads the WebIDL specified in the bindings into an Abstract
Syntax Tree (AST) and performs the mixin operations for implements.

Additionally the specs now use a slightly extended IDL syntax. Instead
of wholesale implementing the second edition of the IDL spec the
parser has been updated to cope with iterator and Promise keywords as
those are the only changes used in the dom and html specifications.

A bug was also fixed in the lexer where negative int literals were
not recognised.