Build interface map allowing for correct dependency generation
This constructs an ordered list of all interfaces in their dependency
order. The topological sort ordering is derived from the interfaces
inheritance.

The resulting table allows the generation phase to easily map
interfaces to classes defined in the binding with a useful ordering.

Additionally it was noticed that the uievent IDL was missing so that
has now been added and allows for a much more complete graph of
interfaces to be constructed.