Add automatic generation of property getters and setters
This allows the binding to omit specifying code for class property
getters and setters. The omitted code will be generated by genbind to
directly call suitable libdom accessors.

The type of the property (string, boolean etc.) is derived from the
IDL. If this type is incorrect, or is a type the generator cannot
automatically produce, then the property will be treated like it has no
binding implementation and generate unimplemented warnings.