restructure AST node creation to avoid casts
This changes Abstract Syntax Tree node creation for both webidl and
genbind syntax tress. If a node is to be created with a numeric value
instead of a pointer a separate API is now used instead of casting
through void.

This fixes parsing and AST building on big endian 64bit platforms
where casting through void, which is completely undefined behaviour,
generates different and non-functioning code. The solution in this
patch is properly portable and correct without relying on casting at
all.

Thanks to James Clarke <jrtc27@jrtc27.com> for the original debugging
and patch demonstrating how to work round the bug.